pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bru-agent"
version = "1.0.0"
description = "BRU - AI agent that connects to Matsya platform"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Matsya", email = "hello@matsyaai.com" }]

dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "anthropic>=0.25.0",
    "httpx>=0.25.0",
    "loguru>=0.7.0",
    "pydantic>=2.5.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
all = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "paramiko>=3.4.0",
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "imapclient>=2.3.1",
    "aiohttp>=3.9.0",
    "cryptography>=41.0.0",
]
browser = [
    "playwright>=1.40.0",
    "selenium>=4.15.0",
    "webdriver-manager>=4.0.0",
]
telegram = ["python-telegram-bot>=21.0"]
social = ["tweepy>=4.14.0"]

[project.scripts]
bru = "bru_agent.cli:main"

[tool.setuptools.packages.find]
include = ["bru_agent*"]

[tool.setuptools.package-data]
bru_agent = ["default_config.yaml"]
